fixed warning

+++ src/utils.hpp
@@ -31,7 +31,7 @@ template <typename T> inline T *safe_mal(unsigned int count) {
 }
 template <typename T> inline std::string vectorString(std::vector<T> vec) {
   std::string res = "[";
-  for (int i = 0; i < vec.size(); i++) {
+  for (size_t i = 0; i < vec.size(); i++) {
     res += std::to_string(vec[i]);
     if (i != vec.size() - 1)
       res += ", ";
@@ -41,7 +41,7 @@ template <typename T> inline std::string vectorString(std::vector<T> vec) {
 template <typename T>
 inline std::string vectorString(std::vector<std::vector<T>> &vec) {
   std::string res = "[";
-  for (int i = 0; i < vec.size(); i++) {
+  for (size_t i = 0; i < vec.size(); i++) {
     res += vectorString(vec[i]);
     if (i != vec.size() - 1)
       res += ",\n";
